MFP2-2886: Adapt gerber-to-svg to its own repo repo (#5)
To work properly, gerber-to-svg must actually be built using webpack, and only then can it be correctly installed by npm from a repo. This PR adapts the tracespace build setup (which relied heavily on things outside the gerber-to-svg directory) to be 100% local to this repo. 

It now supports all npm commands, giving us the ability to build, generate coverage reports, and run 57 unit tests. 

I also added a github workflow to automatically build and test new PRs.  Do note that at least for now, it is important to make sure one runs `npm run build` and then `git add .` and commit and push the built products to the repo as well. 

It also looks like something decided to format everything... this PR does not include any of my changes. The only changes were added files which will show up clean on the diff, and some alterations to package.json.

Ticket: https://macrofab.atlassian.net/browse/MFP2-2886
---------

Co-authored-by: metacollin <metacollin@electropi.mp>